requirements.txt
flask
tflite
nltk
pystempel
https://dl.google.com/coral/python/tflite_runtime-2.1.0.post1-cp36-cp36m-linux_x86_64.whl
gunicorn